+++ src/Cache.php
@@ -93,7 +93,7 @@ use yii\di\Instance;
  * ]);
  * ~~~
  *
- * @property-read bool $isCluster Whether redis is running in cluster mode or not. This property is read-only.
+ * @property-read bool $isCluster Whether redis is running in cluster mode or not.
  *
  * @author Carsten Brandt <mail@cebe.cc>
  * @since 2.0