+++ src/Session.php
@@ -49,7 +49,7 @@ use yii\di\Instance;
  * ]
  * ~~~
  *
- * @property-read bool $useCustomStorage Whether to use custom storage. This property is read-only.
+ * @property-read bool $useCustomStorage Whether to use custom storage.
  *
  * @author Carsten Brandt <mail@cebe.cc>
  * @since 2.0