+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 Yii Framework 2 redis extension Change Log
 ==========================================
 
-2.0.17 under development
-------------------------
+2.0.17 January 11, 2022
+-----------------------
 
 - Enh #176: Fix reconnect logic bug, add `protected function sendRawCommand()` (ilyaplot)